split Context.scala; add deprecation to Has* typeclasses and aliases;
